Rename/move module to effects from side effects
