+++ circuitx/effects/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuitx-effects
+POM_NAME=CircuitX (Effects)
+POM_DESCRIPTION=CircuitX (Effects)
