+++ circuitx/effects/src/androidUnitTest/kotlin/com/slack/circuitx/effects/ImpressionEffectTest.android.kt
@@ -1,0 +1,9 @@
+// Copyright (C) 2023 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuitx.effects
+
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+
+@RunWith(RobolectricTestRunner::class)
+actual class ImpressionEffectTest : ImpressionEffectTestShared by ImpressionEffectTestSharedImpl()
