+++ circuitx/effects/src/jsTest/kotlin/com/slack/circuitx/effects/ImpressionEffectTest.js.kt
@@ -1,0 +1,5 @@
+// Copyright (C) 2023 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuitx.effects
+
+actual class ImpressionEffectTest : ImpressionEffectTestShared by ImpressionEffectTestSharedImpl()
