+++ circuitx/effects/src/jvmTest/kotlin/com/slack/circuitx/effects/ImpressionEffectTest.jvm.kt
@@ -1,0 +1,5 @@
+// Copyright (C) 2023 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuitx.effects
+
+actual class ImpressionEffectTest : ImpressionEffectTestShared by ImpressionEffectTestSharedImpl()
