--- circuitx/sideeffects/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=circuitx-sideeffects
-POM_NAME=CircuitX (Side-effects)
-POM_DESCRIPTION=CircuitX (Side-effects)
