--- circuitx/sideeffects/src/androidUnitTest/kotlin/com/slack/circuitx/sideeffects/ImpressionEffectTest.android.kt
@@ -1,9 +1,0 @@
-// Copyright (C) 2023 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuitx.sideeffects
-
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@RunWith(RobolectricTestRunner::class)
-actual class ImpressionEffectTest : ImpressionEffectTestShared by ImpressionEffectTestSharedImpl()
