--- circuitx/sideeffects/src/iosTest/kotlin/com/slack/circuitx/sideeffects/ImpressionEffectTest.ios.kt
@@ -1,5 +1,0 @@
-// Copyright (C) 2023 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuitx.sideeffects
-
-actual class ImpressionEffectTest : ImpressionEffectTestShared by ImpressionEffectTestSharedImpl()
