+++ settings.gradle.kts
@@ -217,9 +217,9 @@ include(
   ":circuit-runtime-ui",
   ":circuit-test",
   ":circuitx:android",
+  ":circuitx:effects",
   ":circuitx:gesture-navigation",
   ":circuitx:overlays",
-  ":circuitx:sideeffects",
   ":samples:counter",
   ":samples:counter:apps",
   ":samples:counter:mosaic",
